--- students/21K0688/21K0688-p01/src/main/java/ru/mirea/practice/s21K0688/task1/Program1.java
@@ -1,17 +1,0 @@
-package ru.mirea.practice.s21K0688.task1;
-
-import java.util.Scanner;
-
-public class Program1 {
-    public static void main(String[] args) {
-        Scanner sc = new Scanner(System.in);
-        int[] arr = new int[]{1, 5, 13, 2, 6};
-        int sum = 0;
-        for (int i = 0; i < arr.length; i++) {
-            sum += arr[i];
-        }
-        float avg = sum / 5f;
-        System.out.print("Среднее арифметическое = ");
-        System.out.println(avg);
-    }
-}
