--- students/21K0688/21K0688-p01/src/main/java/ru/mirea/practice/s21K0688/task3/Program3.java
@@ -1,9 +1,0 @@
-package ru.mirea.practice.s21K0688.task3;
-
-public class Program3 {
-    public static void main(String[] args) {
-        for (int i = 0; i < args.length; i++) {
-            System.out.println(args[i]);
-        }
-    }
-}
