--- students/21K0688/21K0688-p01/src/main/java/ru/mirea/practice/s21K0688/task4/Program4.java
@@ -1,10 +1,0 @@
-package ru.mirea.practice.s21K0688.task4;
-
-public class Program4 {
-    public static void main(String[] args) {
-        for (int i = 1; i < 11; i++) {
-            double a = 1.0 / i;
-            System.out.printf("%.2f ", a);
-        }
-    }
-}
