+++ students/21K0688/21K0688-p01/src/main/java/ru/mirea/practice/task3/Program3.java
@@ -1,0 +1,9 @@
+package ru.mirea.practice.task3;
+
+abstract class Program3 {
+    public static void main(String[] args) {
+        for (String arg : args) {
+            System.out.println(arg);
+        }
+    }
+}
