+++ students/21K0688/21K0688-p01/src/main/java/ru/mirea/practice/task4/Program4.java
@@ -1,0 +1,10 @@
+package ru.mirea.practice.task4;
+
+abstract class Program4 {
+    public static void main(String[] args) {
+        for (int i = 1; i < 11; i++) {
+            double a = 1.0 / i;
+            System.out.printf("%.2f ", a);
+        }
+    }
+}
