+++ students/21K0688/21K0688-p01/src/main/java/ru/mirea/practice/task5/Factorial.java
@@ -1,0 +1,11 @@
+package ru.mirea.practice.task5;
+
+abstract class Factorial {
+    public static int factor(int a) {
+        int sum = 1;
+        for (int i = 1; i <= a; i++) {
+            sum *= i;
+        }
+        return sum;
+    }
+}
