--- students/21K0688/21K0688-p02/src/main/java/ru/mirea/practice/s21K0688/task2/Ball.java
@@ -1,48 +1,0 @@
-package ru.mirea.practice.s21K0688.task2;
-
-public class Ball {
-    private double x = 0.0;
-    private double y = 0.0;
-
-    public Ball(double x, double y) {
-        this.x = x;
-        this.y = y;
-    }
-
-    public Ball() {
-        this.x = 5;
-        this.y = 5;
-    }
-
-    public double getX() {
-        return x;
-    }
-
-    public void setX(double x) {
-        this.x = x;
-    }
-
-    public double getY() {
-        return y;
-    }
-
-    public void setY(double y) {
-        this.y = y;
-    }
-
-    public void setXY(double x, double y) {
-        this.x = x;
-        this.y = y;
-    }
-
-    public void move(double xDisp, double yDisp) {
-        x += xDisp;
-        y += yDisp;
-    }
-
-    @Override
-    public String toString() {
-        return "x: " + x +
-                ", y: " + y;
-    }
-}
