--- students/21K0688/21K0688-p02/src/main/java/ru/mirea/practice/s21K0688/task3and6/Point.java
@@ -1,32 +1,0 @@
-package ru.mirea.practice.s21K0688.task3and6;
-
-public class Point {
-    private double x;
-    private double y;
-
-    public Point() {
-        x = 0.0;
-        y = 0.0;
-    }
-
-    public Point(double x, double y) {
-        this.x = x;
-        this.y = y;
-    }
-
-    public double getX() {
-        return x;
-    }
-
-    public double getY() {
-        return y;
-    }
-
-    public void setX(double x) {
-        this.x = x;
-    }
-
-    public void setY(double y) {
-        this.y = y;
-    }
-}
