--- students/21K0688/21K0688-p02/src/main/java/ru/mirea/practice/s21K0688/task4/Computer.java
@@ -1,13 +1,0 @@
-package ru.mirea.practice.s21K0688.task4;
-
-public class Computer {
-    private String name;
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getName() {
-        return name;
-    }
-}
