--- students/21K0688/21K0688-p02/src/main/java/ru/mirea/practice/s21K0688/task4/Tester4.java
@@ -1,18 +1,0 @@
-package ru.mirea.practice.s21K0688.task4;
-
-public class Tester4 {
-    private static final int N = 3;
-
-    public static void main(String[] args) {
-        Computer[] arr = new Computer[N];
-        Shop.addComputer(arr);
-        for (int i = 0; i < arr.length; i++) {
-            System.out.print(arr[i].getName() + ", ");
-        }
-        System.out.println("\n");
-        Shop.deleteComputer(arr);
-        for (int i = 0; i < arr.length; i++) {
-            System.out.print(arr[i].getName() + ", ");
-        }
-    }
-}
