--- students/21K0688/21K0688-p02/src/main/java/ru/mirea/practice/s21K0688/task7/Book.java
@@ -1,37 +1,0 @@
-package ru.mirea.practice.s21K0688.task7;
-
-public class Book {
-    private int year;
-    private String author;
-    private String name;
-
-    public int getYear() {
-        return year;
-    }
-
-    public void setYear(int year) {
-        this.year = year;
-    }
-
-    public String getAuthor() {
-        return author;
-    }
-
-    public void setAuthor(String author) {
-        this.author = author;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public Book(String author, int year, String name) {
-        this.author = author;
-        this.name = name;
-        this.year = year;
-    }
-}
