+++ students/21K0688/21K0688-p02/src/main/java/ru/mirea/practice/task1/Author.java
@@ -1,0 +1,36 @@
+package ru.mirea.practice.task1;
+
+public class Author {
+    private String name;
+    private String email;
+    private char gender;
+
+    public Author(String name, String email, char gender) {
+        this.name = name;
+        this.email = email;
+        this.gender = gender;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public char getGender() {
+        return gender;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    @Override
+    public String toString() {
+        return "name: " + name
+                + ", email: " + email
+                + ", gender: " + gender;
+    }
+}
