+++ students/21K0688/21K0688-p02/src/main/java/ru/mirea/practice/task4/Computer.java
@@ -1,0 +1,13 @@
+package ru.mirea.practice.task4;
+
+public class Computer {
+    private String name;
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
