+++ students/21K0688/21K0688-p02/src/main/java/ru/mirea/practice/task4/Tester4.java
@@ -1,0 +1,18 @@
+package ru.mirea.practice.task4;
+
+abstract class Tester4 {
+    private static final int N = 3;
+
+    public static void main(String[] args) {
+        Computer[] arr = new Computer[N];
+        Shop.addComputer(arr);
+        for (Computer computer : arr) {
+            System.out.print(computer.getName() + ", ");
+        }
+        System.out.println("\n");
+        Shop.deleteComputer(arr);
+        for (Computer computer : arr) {
+            System.out.print(computer.getName() + ", ");
+        }
+    }
+}
