+++ students/21K0688/21K0688-p02/src/main/java/ru/mirea/practice/task8/Tester8.java
@@ -1,0 +1,22 @@
+package ru.mirea.practice.task8;
+
+abstract class Tester8 {
+    public static void main(String[] args) {
+        String[] arr = {"a", "b", "c", "d", "f"};
+        String temp;
+        int n = arr.length;
+        for (String s : arr) {
+            System.out.print(s + " ");
+        }
+        System.out.print("\n");
+        for (int i = 0; i < arr.length / 2; i++) {
+            temp = arr[i];
+            arr[i] = arr[n - 1];
+            arr[n - 1] = temp;
+            n--;
+        }
+        for (String s : arr) {
+            System.out.print(s + " ");
+        }
+    }
+}
